Updated Quiz class
Created quiz class for easy acces for the GUI.
