+++ build.xml
@@ -7,9 +7,9 @@
 <!-- the Compile on Save feature is turned off for the project. -->
 <!-- You can turn off the Compile on Save (or Deploy on Save) setting -->
 <!-- in the project's Project Properties dialog box.-->
-<project name="com.quiz.quiz.Quiz" default="default" basedir=".">
+<project name="com.quiz.quiz.QuizConsole" default="default" basedir=".">
     <target name="default"/>
-    <description>Builds, tests, and runs the project com.quiz.quiz.Quiz.</description>
+    <description>Builds, tests, and runs the project com.quiz.quiz.QuizConsole.</description>
     <import file="nbproject/build-impl.xml"/>
     <!--
 
@@ -59,7 +59,7 @@
 
     An example of overriding the target for project execution could look like this:
 
-        <target name="run" depends="Quiz-impl.jar">
+        <target name="run" depends="QuizConsole-impl.jar">
             <exec dir="bin" executable="launcher.exe">
                 <arg file="${dist.jar}"/>
             </exec>
